Adding password to sync page
Closes #738
