+++ src/ui/views/Setting/Backups/BackupsPassword.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import SettingsPassword from '@/ui/FRWComponent/SettingsPassword';
+
+const BackupsPassword = () => {
+  return <SettingsPassword verifiedUrl="/dashboard/setting/backups"></SettingsPassword>;
+};
+
+export default BackupsPassword;
